requirements.txt
docxtpl
python-docx
pyyaml
